Create README.md
This folder consists of the matlab codes used for the optimal control of 'CTCR' robots. Open the matlab and run the command 'optimize_along_path([])' to generate the optimal solutions (sol). Run the command 'Animate_lefthanded(sol, F)', to visualize the evolution of the CTCR configurations and control parameters. Here, $F$ is the tip Load used for the optimization problem